+++ src/main/java/com/technova/technov/domain/repository/MensajeEmpleadoRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MensajeEmpleadoRepository extends JpaRepository<MensajeEmpleado, Long> {
-    List<MensajeEmpleado> findByEmpleadoIdOrderByCreatedAtDesc(Long empleadoId);
-    List<MensajeEmpleado> findByEmpleadoIdAndLeidoOrderByCreatedAtDesc(Long empleadoId, boolean leido);
+public interface MensajeEmpleadoRepository extends JpaRepository<MensajeEmpleado, Integer> {
+    List<MensajeEmpleado> findByEmpleadoIdOrderByCreatedAtDesc(Integer empleadoId);
+    List<MensajeEmpleado> findByEmpleadoIdAndLeidoOrderByCreatedAtDesc(Integer empleadoId, boolean leido);
     List<MensajeEmpleado> findByTipoAndPrioridadOrderByCreatedAtDesc(String tipo, String prioridad);
 }
